code/CloudStorageTrigger/requirements.txt
wand
google-cloud-storage
